# Speed up Docker image creation