--- a/Dockerfile
+++ b/Dockerfile
@@ -1,5 +1,5 @@
 ################################################################################
-# Build `noseyparker`
+# Build an image used for all building actions
 #
 # We use the oldest Debian-based image that can build Nosey Parker without trouble.
 # This is done in an effort to link against an older glibc, so that the built
@@ -9,7 +9,27 @@
 #
 # See https://github.com/praetorian-inc/noseyparker/issues/58.
 ################################################################################
-FROM rust:1.76-bullseye AS builder
+FROM rust:1.76-bullseye AS chef 
+# We only pay the installation cost once, 
+# it will be cached from the second build onwards
+RUN cargo install cargo-chef 
+
+WORKDIR "/noseyparker"
+
+################################################################################
+# Generate a `recipe.json` file to capture the set of information required to
+# build the dependencies of `noseyparker`.
+################################################################################
+FROM chef AS planner
+
+COPY . .
+
+RUN cargo chef prepare  --recipe-path recipe.json
+
+################################################################################
+# Build `noseyparker`
+################################################################################
+FROM chef AS builder
 
 # Install dependencies
 #
@@ -22,7 +42,11 @@ RUN apt-get update &&\
         &&\
     apt-get clean
 
-WORKDIR "/noseyparker"
+COPY --from=planner /noseyparker/recipe.json recipe.json
+
+# Build dependencies - this is the caching Docker layer
+# Arguments match arguments specified in `create-release.zsh` script
+RUN cargo chef cook --locked --profile release --features "release" --recipe-path recipe.json
 
 COPY . .